# prodsys_openAPI.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "prodsys API",
    "description": "\nThe prodsys API allows you to create and run production simulations and optimizations with the prodsys library as a web service. \n",
    "contact": {
      "name": "Sebastian Behrendt",
      "email": "sebastian.behrendt@kit.edu"
    },
    "license": {
      "name": "MIT License",
      "url": "https://mit-license.org/"
    },
    "version": "0.4.2"
  },
  "paths": {
...
    "/projects/{project_id}/adapters/{adapter_id}/results/static_results": {
      "get": {
        "tags": [
          "results"
        ],
        "summary": "Get Static Results",
        "operationId": "get_static_results_projects__project_id__adapters__adapter_id__results_static_results_get",
        "parameters": [
          {
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            },
            "name": "project_id",
            "in": "path"
          },
          {
            "required": true,
            "schema": {
              "type": "string",
              "title": "Adapter Id"
            },
            "name": "adapter_id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Sucessfully returned static results",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "anyOf": [
                      {
                        "$ref": "#/components/schemas/Output"
                      },
                      {
                        "$ref": "#/components/schemas/Throughput"
                      },
                      {
                        "$ref": "#/components/schemas/Cost"
                      },
                      {
                        "$ref": "#/components/schemas/WIP"
                      },
                      {
                        "$ref": "#/components/schemas/ThroughputTime"
                      },
                      {
                        "$ref": "#/components/schemas/ProcessingTime"
                      },
                      {
                        "$ref": "#/components/schemas/ProductiveTime"
                      },
                      {
                        "$ref": "#/components/schemas/StandbyTime"
                      },
                      {
                        "$ref": "#/components/schemas/SetupTime"
                      },
                      {
                        "$ref": "#/components/schemas/UnscheduledDowntime"
                      },
                      {
                        "$ref": "#/components/schemas/DynamicWIP"
                      },
                      {
                        "$ref": "#/components/schemas/DynamicThroughputTime"
                      }
                    ]
                  },
                  "type": "array",
                  "title": "Response Get Static Results Projects  Project Id  Adapters  Adapter Id  Results Static Results Get"
                },
                "example": [
                  {
                    "name": "output",
                    "target": "max",
                    "weight": 1,
                    "value": 34,
                    "context": [
                      "system",
                      "product_type"
                    ],
                    "product_type": "ProductType_1"
                  },
                  {
                    "name": "throughput",
                    "target": "max",
                    "weight": 1,
                    "value": 4.32,
                    "context": [
                      "system",
                      "product_type"
                    ],
                    "product_type": "ProductType_1"
                  },
                  {
                    "name": "cost",
                    "target": "min",
                    "weight": 0.5,
                    "value": 36000,
                    "context": [
                      "system"
                    ]
                  },
                  {
                    "name": "WIP",
                    "target": "min",
                    "weight": 1,
                    "value": 121,
                    "context": [
                      "system",
                      "product_type"
                    ],
                    "product_type": "ProductType_1"
                  },
                  {
                    "name": "throughput_time",
                    "target": "min",
                    "weight": 1,
                    "value": 221.1,
                    "context": [
                      "system",
                      "product_type"
                    ],
                    "product_type": "ProductType_1"
                  },
                  {
                    "name": "processing_time",
                    "target": "min",
                    "weight": 1,
                    "value": 1.2,
                    "context": [
                      "resource",
                      "process"
                    ],
                    "resource": "Resource_1",
                    "process": "P1"
                  },
                  {
                    "name": "productive_time",
                    "target": "max",
                    "weight": 1,
                    "value": 0.65,
                    "context": [
                      "resource"
                    ],
                    "resource": "Resource_1"
                  },
                  {
                    "name": "standby_time",
                    "target": "min",
                    "weight": 1,
                    "value": 0.12,
                    "context": [
                      "resource"
                    ],
                    "resource": "Resource_1"
                  },
                  {
                    "name": "setup_time",
                    "target": "min",
                    "weight": 1,
                    "value": 0.08,
                    "context": [
                      "resource"
                    ],
                    "resource": "Resource_1"
                  },
                  {
                    "name": "unscheduled_downtime",
                    "target": "min",
                    "weight": 1,
                    "value": 0.1,
                    "context": [
                      "resource"
                    ],
                    "resource": "Resource_1"
                  },
                  {
                    "name": "dynamic_WIP",
                    "target": "min",
                    "weight": 1,
                    "value": 121,
                    "context": [
                      "system",
                      "product"
                    ],
                    "product_type": "ProductType_1",
                    "start_time": 21.2,
                    "end_time": 23.4
                  },
                  {
                    "name": "throughput_time",
                    "target": "min",
                    "weight": 1,
                    "value": 201.3,
                    "context": [
                      "system",
                      "product"
                    ],
                    "product_type": "ProductType_1",
                    "product": "Product_1_23"
                  }
                ]
              }
            }
          },
          "404": {
            "description": "No KPIs found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/projects/{project_id}/adapters/{adapter_id}/results/event_results": {
      "get": {
        "tags": [
          "results"
        ],
        "summary": "Get Event Results",
        "operationId": "get_event_results_projects__project_id__adapters__adapter_id__results_event_results_get",
        "parameters": [
          {
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            },
            "name": "project_id",
            "in": "path"
          },
          {
            "required": true,
            "schema": {
              "type": "string",
              "title": "Adapter Id"
            },
            "name": "adapter_id",
            "in": "path"
          },
          {
            "required": false,
            "schema": {
              "type": "integer",
              "title": "Num Events",
              "default": 10
            },
            "name": "num_events",
            "in": "query"
          },
          {
            "required": false,
            "schema": {
              "type": "integer",
              "title": "Offset",
              "default": 0
            },
            "name": "offset",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Sucessfully returned event results",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Event"
                  },
                  "type": "array",
                  "title": "Response Get Event Results Projects  Project Id  Adapters  Adapter Id  Results Event Results Get"
                },
                "example": [
                  {
                    "time": 12,
                    "resource": "R1",
                    "state": "P1",
                    "state_type": "Production",
                    "activity": "start state",
                    "product": "Product_1_12",
                    "expected_end_time": 24.3
                  },
                  {
                    "time": 24.3,
                    "resource": "R1",
                    "state": "P1",
                    "state_type": "Production",
                    "activity": "end state",
                    "product": "Product_1_12"
                  }
                ]
              }
            }
          },
          "404": {
            "description": "No events found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/projects/{project_id}/adapters/{adapter_id}/results/{kpi}": {
      "get": {
        "tags": [
          "results"
        ],
        "summary": "Get Kpi",
        "operationId": "get_kpi_projects__project_id__adapters__adapter_id__results__kpi__get",
        "parameters": [
          {
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            },
            "name": "project_id",
            "in": "path"
          },
          {
            "required": true,
            "schema": {
              "type": "string",
              "title": "Adapter Id"
            },
            "name": "adapter_id",
            "in": "path"
          },
          {
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/KPIEnum"
            },
            "name": "kpi",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Sucessfully returned static results",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "anyOf": [
                      {
                        "$ref": "#/components/schemas/Output"
                      },
                      {
                        "$ref": "#/components/schemas/Throughput"
                      },
                      {
                        "$ref": "#/components/schemas/Cost"
                      },
                      {
                        "$ref": "#/components/schemas/WIP"
                      },
                      {
                        "$ref": "#/components/schemas/ThroughputTime"
                      },
                      {
                        "$ref": "#/components/schemas/ProcessingTime"
                      },
                      {
                        "$ref": "#/components/schemas/ProductiveTime"
                      },
                      {
                        "$ref": "#/components/schemas/StandbyTime"
                      },
                      {
                        "$ref": "#/components/schemas/SetupTime"
                      },
                      {
                        "$ref": "#/components/schemas/UnscheduledDowntime"
                      },
                      {
                        "$ref": "#/components/schemas/DynamicWIP"
                      },
                      {
                        "$ref": "#/components/schemas/DynamicThroughputTime"
                      }
                    ]
                  },
                  "type": "array",
                  "title": "Response Get Kpi Projects  Project Id  Adapters  Adapter Id  Results  Kpi  Get"
                },
                "examples": {
                  "Output": {
                    "summary": "Output KPI",
                    "value": {
                      "name": "output",
                      "target": "max",
                      "weight": 1,
                      "value": 34,
                      "context": [
                        "system",
                        "product_type"
                      ],
                      "product_type": "ProductType_1"
                    }
                  },
                  "Throughput": {
                    "summary": "Throughput KPI",
                    "value": {
                      "name": "throughput",
                      "target": "max",
                      "weight": 1,
                      "value": 4.32,
                      "context": [
                        "system",
                        "product_type"
                      ],
                      "product_type": "ProductType_1"
                    }
                  },
                  "Cost": {
                    "summary": "Cost KPI",
                    "value": {
                      "name": "cost",
                      "target": "min",
                      "weight": 0.5,
                      "value": 36000,
                      "context": [
                        "system"
                      ]
                    }
                  },
                  "WIP": {
                    "summary": "WIP KPI",
                    "value": {
                      "name": "WIP",
                      "target": "min",
                      "weight": 1,
                      "value": 121,
                      "context": [
                        "system",
                        "product_type"
                      ],
                      "product_type": "ProductType_1"
                    }
                  },
                  "ThroughputTime": {
                    "summary": "Throughput time KPI",
                    "value": {
                      "name": "throughput_time",
                      "target": "min",
                      "weight": 1,
                      "value": 221.1,
                      "context": [
                        "system",
                        "product_type"
                      ],
                      "product_type": "ProductType_1"
                    }
                  },
                  "ProcessingTime": {
                    "summary": "Processing time KPI",
                    "value": {
                      "name": "processing_time",
                      "target": "min",
                      "weight": 1,
                      "value": 1.2,
                      "context": [
                        "resource",
                        "process"
                      ],
                      "resource": "Resource_1",
                      "process": "P1"
                    }
                  },
                  "ProductiveTime": {
                    "summary": "Productive time KPI",
                    "value": {
                      "name": "productive_time",
                      "target": "max",
                      "weight": 1,
                      "value": 0.65,
                      "context": [
                        "resource"
                      ],
                      "resource": "Resource_1"
                    }
                  },
                  "StandbyTime": {
                    "summary": "Standby time KPI",
                    "value": {
                      "name": "standby_time",
                      "target": "min",
                      "weight": 1,
                      "value": 0.12,
                      "context": [
                        "resource"
                      ],
                      "resource": "Resource_1"
                    }
                  },
                  "SetupTime": {
                    "summary": "Setup time KPI",
                    "value": {
                      "name": "setup_time",
                      "target": "min",
                      "weight": 1,
                      "value": 0.08,
                      "context": [
                        "resource"
                      ],
                      "resource": "Resource_1"
                    }
                  },
                  "UnscheduledDowntime": {
                    "summary": "Unscheduled downtime KPI",
                    "value": {
                      "name": "unscheduled_downtime",
                      "target": "min",
                      "weight": 1,
                      "value": 0.1,
                      "context": [
                        "resource"
                      ],
                      "resource": "Resource_1"
                    }
                  },
                  "DynamicWIP": {
                    "summary": "Dynamic WIP KPI",
                    "value": {
                      "name": "dynamic_WIP",
                      "target": "min",
                      "weight": 1,
                      "value": 121,
                      "context": [
                        "system",
                        "product"
                      ],
                      "product_type": "ProductType_1",
                      "start_time": 21.2,
                      "end_time": 23.4
                    }
                  },
                  "DynamicThroughputTime": {
                    "summary": "Dynamic throughput time KPI",
                    "value": {
                      "name": "throughput_time",
                      "target": "min",
                      "weight": 1,
                      "value": 201.3,
                      "context": [
                        "system",
                        "product"
                      ],
                      "product_type": "ProductType_1",
                      "product": "Product_1_23"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "No KPIs found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Cost": {
        "properties": {
          "name": {
            "type": "string",
            "enum": [
              "cost"
            ],
            "title": "Name"
          },
          "target": {
            "type": "string",
            "enum": [
              "min"
            ],
            "title": "Target",
            "default": "min"
          },
          "weight": {
            "type": "number",
            "title": "Weight",
            "default": 1
          },
          "value": {
            "type": "number",
            "title": "Value"
          },
          "context": {
            "items": {
              "$ref": "#/components/schemas/KPILevelEnum"
            },
            "type": "array"
          },
          "resource": {
            "type": "string",
            "title": "Resource"
          },
          "product_type": {
            "type": "string",
            "title": "Product Type"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "Cost",
        "description": "Class that represents a KPI. Not intended for usage but only inheritance.\n\nArgs:\n    name (KPIEnum): Name of the KPI.\n    target (Literal[\"min\", \"max\"]): Favourable target of the KPI.\n    weight (Optional[float], optional): Weight of the KPI. Defaults to 1.\n    value (Optional[float], optional): Value of the KPI. Defaults to None.\n    context (Tuple[KPILevelEnum, ...], optional): Context of the KPI. Defaults to None.\n    resource (Optional[str], optional): Resource of the KPI. Defaults to None.\n    product_type (Optional[str], optional): Product type of the KPI. Defaults to None.    Returns:",
        "example": {
          "summary": "Cost KPI",
          "value": {
            "name": "cost",
            "target": "min",
            "weight": 0.5,
            "value": 36000,
            "context": [
              "system"
            ]
          }
        }
      },
...
      "Output": {
        "properties": {
          "name": {
            "type": "string",
            "enum": [
              "output"
            ],
            "title": "Name",
            "default": "output"
          },
          "target": {
            "type": "string",
            "enum": [
              "max"
            ],
            "title": "Target",
            "default": "max"
          },
          "weight": {
            "type": "number",
            "title": "Weight",
            "default": 1
          },
          "value": {
            "type": "number",
            "title": "Value"
          },
          "context": {
            "items": {
              "$ref": "#/components/schemas/KPILevelEnum"
            },
            "type": "array"
          },
          "resource": {
            "type": "string",
            "title": "Resource"
          },
          "product_type": {
            "type": "string",
            "title": "Product Type"
          }
        },
        "type": "object",
        "title": "Output",
        "description": "Class that represents a KPI. Not intended for usage but only inheritance.\n\nArgs:\n    name (KPIEnum): Name of the KPI.\n    target (Literal[\"min\", \"max\"]): Favourable target of the KPI.\n    weight (Optional[float], optional): Weight of the KPI. Defaults to 1.\n    value (Optional[float], optional): Value of the KPI. Defaults to None.\n    context (Tuple[KPILevelEnum, ...], optional): Context of the KPI. Defaults to None.\n    resource (Optional[str], optional): Resource of the KPI. Defaults to None.\n    product_type (Optional[str], optional): Product type of the KPI. Defaults to None.    Returns:",
        "example": {
          "summary": "Output KPI",
          "value": {
            "name": "output",
            "target": "max",
            "weight": 1,
            "value": 34,
            "context": [
              "system",
              "product_type"
            ],
            "product_type": "ProductType_1"
          }
        }
      },
...
      "ProductiveTime": {
        "properties": {
          "name": {
            "type": "string",
            "enum": [
              "productive_time"
            ],
            "title": "Name"
          },
          "target": {
            "type": "string",
            "enum": [
              "max"
            ],
            "title": "Target",
            "default": "max"
          },
          "weight": {
            "type": "number",
            "title": "Weight",
            "default": 1
          },
          "value": {
            "type": "number",
            "title": "Value"
          },
          "context": {
            "items": {
              "$ref": "#/components/schemas/KPILevelEnum"
            },
            "type": "array"
          },
          "resource": {
            "type": "string",
            "title": "Resource"
          },
          "product_type": {
            "type": "string",
            "title": "Product Type"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "ProductiveTime",
        "description": "Class that represents a KPI. Not intended for usage but only inheritance.\n\nArgs:\n    name (KPIEnum): Name of the KPI.\n    target (Literal[\"min\", \"max\"]): Favourable target of the KPI.\n    weight (Optional[float], optional): Weight of the KPI. Defaults to 1.\n    value (Optional[float], optional): Value of the KPI. Defaults to None.\n    context (Tuple[KPILevelEnum, ...], optional): Context of the KPI. Defaults to None.\n    resource (Optional[str], optional): Resource of the KPI. Defaults to None.\n    product_type (Optional[str], optional): Product type of the KPI. Defaults to None.    Returns:",
        "example": {
          "summary": "Productive time KPI",
          "value": {
            "name": "productive_time",
            "target": "max",
            "weight": 1,
            "value": 0.65,
            "context": [
              "resource"
            ],
            "resource": "Resource_1"
          }
        }
      },
...
      "Throughput": {
        "properties": {
          "name": {
            "type": "string",
            "enum": [
              "throughput"
            ],
            "title": "Name"
          },
          "target": {
            "type": "string",
            "enum": [
              "max"
            ],
            "title": "Target",
            "default": "max"
          },
          "weight": {
            "type": "number",
            "title": "Weight",
            "default": 1
          },
          "value": {
            "type": "number",
            "title": "Value"
          },
          "context": {
            "items": {
              "$ref": "#/components/schemas/KPILevelEnum"
            },
            "type": "array"
          },
          "resource": {
            "type": "string",
            "title": "Resource"
          },
          "product_type": {
            "type": "string",
            "title": "Product Type"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "Throughput",
        "description": "Class that represents a KPI. Not intended for usage but only inheritance.\n\nArgs:\n    name (KPIEnum): Name of the KPI.\n    target (Literal[\"min\", \"max\"]): Favourable target of the KPI.\n    weight (Optional[float], optional): Weight of the KPI. Defaults to 1.\n    value (Optional[float], optional): Value of the KPI. Defaults to None.\n    context (Tuple[KPILevelEnum, ...], optional): Context of the KPI. Defaults to None.\n    resource (Optional[str], optional): Resource of the KPI. Defaults to None.\n    product_type (Optional[str], optional): Product type of the KPI. Defaults to None.    Returns:",
        "example": {
          "summary": "Throughput KPI",
          "value": {
            "name": "throughput",
            "target": "max",
            "weight": 1,
            "value": 4.32,
            "context": [
              "system",
              "product_type"
            ],
            "product_type": "ProductType_1"
          }
        }
      },
      "ThroughputTime": {
        "properties": {
          "name": {
            "type": "string",
            "enum": [
              "throughput_time"
            ],
            "title": "Name"
          },
          "target": {
            "type": "string",
            "enum": [
              "min"
            ],
            "title": "Target",
            "default": "min"
          },
          "weight": {
            "type": "number",
            "title": "Weight",
            "default": 1
          },
          "value": {
            "type": "number",
            "title": "Value"
          },
          "context": {
            "items": {
              "$ref": "#/components/schemas/KPILevelEnum"
            },
            "type": "array"
          },
          "resource": {
            "type": "string",
            "title": "Resource"
          },
          "product_type": {
            "type": "string",
            "title": "Product Type"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "ThroughputTime",
        "description": "Class that represents a KPI. Not intended for usage but only inheritance.\n\nArgs:\n    name (KPIEnum): Name of the KPI.\n    target (Literal[\"min\", \"max\"]): Favourable target of the KPI.\n    weight (Optional[float], optional): Weight of the KPI. Defaults to 1.\n    value (Optional[float], optional): Value of the KPI. Defaults to None.\n    context (Tuple[KPILevelEnum, ...], optional): Context of the KPI. Defaults to None.\n    resource (Optional[str], optional): Resource of the KPI. Defaults to None.\n    product_type (Optional[str], optional): Product type of the KPI. Defaults to None.    Returns:",
        "example": {
          "summary": "Throughput time KPI",
          "value": {
            "name": "throughput_time",
            "target": "min",
            "weight": 1,
            "value": 221.1,
            "context": [
              "system",
              "product_type"
            ],
            "product_type": "ProductType_1"
          }
        }
      },
...
      "WIP": {
        "properties": {
          "name": {
            "type": "string",
            "enum": [
              "WIP"
            ],
            "title": "Name"
          },
          "target": {
            "type": "string",
            "enum": [
              "min"
            ],
            "title": "Target",
            "default": "min"
          },
          "weight": {
            "type": "number",
            "title": "Weight",
            "default": 1
          },
          "value": {
            "type": "number",
            "title": "Value"
          },
          "context": {
            "items": {
              "$ref": "#/components/schemas/KPILevelEnum"
            },
            "type": "array"
          },
          "resource": {
            "type": "string",
            "title": "Resource"
          },
          "product_type": {
            "type": "string",
            "title": "Product Type"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "WIP",
        "description": "Class that represents a KPI. Not intended for usage but only inheritance.\n\nArgs:\n    name (KPIEnum): Name of the KPI.\n    target (Literal[\"min\", \"max\"]): Favourable target of the KPI.\n    weight (Optional[float], optional): Weight of the KPI. Defaults to 1.\n    value (Optional[float], optional): Value of the KPI. Defaults to None.\n    context (Tuple[KPILevelEnum, ...], optional): Context of the KPI. Defaults to None.\n    resource (Optional[str], optional): Resource of the KPI. Defaults to None.\n    product_type (Optional[str], optional): Product type of the KPI. Defaults to None.    Returns:",
        "example": {
          "summary": "WIP KPI",
          "value": {
            "name": "WIP",
            "target": "min",
            "weight": 1,
            "value": 121,
            "context": [
              "system",
              "product_type"
            ],
            "product_type": "ProductType_1"
          }
        }
      }
    }
  }
}